+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 
 ## 0.0.28
 
-
+- Update default runtime version to v1.7.0
 
 ## 0.0.27